move photo from public file
