+++ client/src/components/sections/whoAreWe/style.css
